Move binaries only if they exist

+++ Lab 2/compile.sh
@@ -4,6 +4,12 @@ ghc "$DIR/client.hs"
 ghc "$DIR/server.hs"
 rm -rf "$DIR/bin"
 mkdir -p "$DIR/bin"
-mv "$DIR/client" "$DIR/bin"
-mv "$DIR/server" "$DIR/bin"
+if [ -f "$DIR/client" ]
+then
+    mv "$DIR/client" "$DIR/bin"
+fi
+if [ -f "$DIR/server" ]
+then
+    mv "$DIR/server" "$DIR/bin"
+fi
 find "$DIR" -name "*.o" -o -name "*.hi" | xargs -I{} rm {}
